+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - This file!
 - Support for placeholders in flag usage strings
 - `App.Metadata` map for arbitrary data/state management
+- `Set` and `GlobalSet` methods on `*cli.Context` for altering values after
+parsing.
 
 ### Changed
 - The `App.Action` and `Command.Action` now prefer a return signature of